Return correct costs from single_source_shortest_paths()
Return the *computed* cost for traversing the edge as part of the
predecessor map.

No longer return the {v => cost from s to v} map, which wasn't being
used anyway (at least not internally).

Previously, extract_shortest_path_from_predecessor_list() used the first
item of the edge attribute list as the cost for crossing an edge. Now it
uses the computed cost.