Simplify graph structure
The structure is now a simple adjacency matrix::

    {node: {neighbor_node: edge, ...}, ...}

Before, the structure was a bit more complicated--the adjacency matrix
contained pointers to edges in a separate dict instead of the edges
themselves.

The nodes and edges can be any type of object as long as the nodes are
hashable, because they are used as dict keys.

Added a simple Graph class to encapsulate this.

single_source_shortest_paths() also now makes less assumptions about
the types of the nodes and edges a graph contains. In particular, it's
no longer assumed that edges are lists containing a cost as their first
item.