Return graph loaded by pickle.load directly
Before, the graph returned from pickle.load was being used to create
a new Graph, but this is unnecessary because pickle.load returns a fully
formed, ready to use Graph instance.

Interestingly, this has a negligible impact on performance (as measured
by timeit.timeit)--the call to dict.update in Graph.__init__ took
essentially no time.