Add Graph.remove_node()
This moves the logic for removing a node from `Graph.__del__()` to
`Graph.remove_node()`. The new version not only removes the node from
the underlying data dict, which has the effect of removing outgoing
edges, it also removes the node's incoming edges instead of letting them
dangle in the graph.