Add neighboring nodes when adding a node or edge
Added tests to exercise most directed graph functionality. These tests
revealed some issues, in particular that neighboring nodes weren't being
added in some cases where they might logically be expected to be added.

- When adding a node with neighbors, add a node for each neighbor that
  isn't already present in the graph.
- Likewise, when adding an edge `(u, v)`, add a node for `v` when it's
  not already present in the graph.

This is being added to v3 because it changes the underlying graph
structure in a potentially backward-incompatible way. It also requires
a bit of extra checking when an annex graph is passed to
`single_source_shortest_paths()`.